# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel9 APP API",
        "version": "V1"
    },
    "servers": [
        {
            "url": "http://localhost/laravel9/public/api/v1"
        }
    ],
    "paths": {
...
        "/user/details": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User Details",
                "description": "",
                "operationId": "User Details",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/updateprofile": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update Profile",
                "description": "Update User Profile",
                "operationId": "updateUserProfile",
                "requestBody": {
                    "description": "User Detail",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Laravel Admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "laraveladmin@yopmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry,Some thing went wrong. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/change-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change Password",
                "description": "Change Password",
                "operationId": "changePassword",
                "requestBody": {
                    "description": "Change Password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "currentpassword",
                                    "password",
                                    "confirmpassword"
                                ],
                                "properties": {
                                    "currentpassword": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Moon@123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Moon@123"
                                    },
                                    "confirmpassword": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Moon@123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    }
}